amazon: Allow importing separately-charged orders
I believe a single Amazon shipment consistently applies to a particular
credit card charge, and a single credit card charge consistently applies
to a particular Amazon order. However, Amazon sometimes charges credit
cards multiple times for a single order and a single credit card charge
can sometimes reflect multiple shipments.

With this change, if we don't match an order one to one with a charge,
we try different combinations of line items and see if those match all
the charges.